Never use HTML feeds if developer mode is not on
Specifically one of my subscribed feeds introduced a captcha to their
feed (probably unintentional, but it happens). This caused Gatherss to
interpret it as an HTML feed, since its initial content type was HTML.
Probably if developer mode is not on, it should never even try to use
the HTML feed logic.