Better (harder) refreshing for individual feeds
Previously icons and other data were not updated even when using the
refresh button for an individual feed. To _really_ update it, one had to
remove it entirely and re-add it. Now, this is the internal behavior;
when refreshing an individual feed, it is entirely remove and re-added,
causing icons and other metadata to be re-fetched properly.